Prune resources with ownerReferences if they belong to Stack
When eg. a controller is the owner of its resources, it also controls
the metadata.ownerReferences field. Stack is not able to prune its
resources correctly if any resources with ownerReferences are kept.

When passing owner_reference: to Stack#apply or Stack#prune, any
owned resources matching to given owner_reference will be deemed
ours and be pruned.